gen.py:Fix wrong gen_files_conf generation
When a VM is declared in a shim DIF but not in a normal DIF the generation of
gen_files_conf fails because gen.py always adds to the list files in the form
of normal.%(name)s.*.dif. This makes the scp command failed and the up.sh gets
stuck in an infinite loop. This commits checks that the VM is part of the
normal DIF before adding the corresponding files.